+++ Gemfile
@@ -37,6 +37,13 @@ gem 'jbuilder', '~> 2.5'
 gem 'bootsnap', '>= 1.1.0', require: false
 
 group :development, :test do
+  gem 'rspec-rails'
+
+    gem "factory_bot_rails"
+    gem 'rspec-parameterized'
+    gem 'faker', '~> 2.0.0'
+    gem 'shoulda-matchers', require: false
+    gem 'rails-controller-testing'
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
   gem 'byebug', platforms: [:mri, :mingw, :x64_mingw]
 
@@ -54,16 +61,11 @@ end
 group :test do
   # Adds support for Capybara system testing and selenium driver
   gem 'capybara', '>= 2.15'
-  gem 'selenium-webdriver'
+  # gem 'selenium-webdriver'
   # Easy installation and use of chromedriver to run system tests with Chrome
-  gem 'chromedriver-helper'
-  gem 'rspec-rails'
-
-  gem "factory_bot_rails"
-    gem 'rspec-parameterized'
-    gem 'faker', '~> 2.0.0'
-    gem 'shoulda-matchers', require: false
-    gem 'rails-controller-testing'
+  # gem 'chromedriver-helper'
+  gem 'webdrivers'
+  
 end
