+++ app/models/end_user.rb
@@ -2,6 +2,9 @@ class EndUser < ApplicationRecord
   acts_as_paranoid
   # Include default devise modules. Others available are:
   # :confirmable, :lockable, :timeoutable, :trackable and :omniauthable
+  # validates :first_name, presence: true
+  # validates :last_name, presence: true
+  
   devise :database_authenticatable, :registerable,
         :recoverable, :rememberable, :validatable
   has_many :addresses, dependent: :destroy
