+++ db/schema.rb
@@ -94,11 +94,6 @@ ActiveRecord::Schema.define(version: 2019_10_21_072517) do
     t.datetime "updated_at", null: false
   end
 
-  create_table "homes", force: :cascade do |t|
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
-  end
-
   create_table "labels", force: :cascade do |t|
     t.string "label_name"
     t.datetime "created_at", null: false
