+++ spec/rails_helper.rb
@@ -7,7 +7,8 @@ require File.expand_path('../config/environment', __dir__)
 # Prevent database truncation if the environment is production
 abort("The Rails environment is running in production mode!") if Rails.env.production?
 require 'rspec/rails'
-require 'shoulda/matchers'
+require 'devise'
+require 'shoulda-matchers'
 
 # Add additional requires below this line. Rails is not loaded until this point!
